agent_framework/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-socketio==5.11.0
pyautogui==0.9.54
pynacl==1.5.0
argon2-cffi==23.1.0
faiss-gpu==1.8.0
voyageai==0.2.3
fastmcp==0.4.2
openai==1.14.3
fast-agent==0.3.1
rich==13.7.1
gitpython==3.1.43
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1 